Updated unit test file names to match Maven conventions
